Merge with old context instead of overriding it
Sending a PR because it's probably worth a thousand word. Does this make sense? I make my own additions to the pages' `context` and I found that this plugin is overriding those. It could be that I'm just getting not Gatsby right and there's something weird with my setup though.